PSUSensor: Fix error identification for scalefactor smaller than 1 cases
To fix cases that scalefactor < 1,
revise scalefactor variable type from unsigned int to double.

Tested:
Get adc128d818 and ina233 reading from PSUSensor by dbus introspect

+++ src/PSUSensor.cpp
@@ -42,7 +42,7 @@ PSUSensor::PSUSensor(const std::string& path, const std::string& objectType,
                      std::vector<thresholds::Threshold>&& thresholdsIn,
                      const std::string& sensorConfiguration,
                      const PowerState& powerState,
-                     const std::string& sensorUnits, unsigned int factor,
+                     const std::string& sensorUnits, double factor,
                      double max, double min, double offset,
                      const std::string& label, size_t tSize, double pollRate,
                      const std::shared_ptr<I2CDevice>& i2cDevice) :
